third_party/blink/web_tests/platform/win/external/wpt/url/a-element-origin-expected.txt

This is a testharness.js-based test.
Found 11 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Parsing origin: <a:	 foo.com> against <http://example.org/foo/bar>
  assert_equals: origin expected "null" but got "file://"
[FAIL] Parsing origin: <\\\\x\\hello> against <http://example.org/foo/bar>
  assert_equals: origin expected "http://x" but got "file://"
[FAIL] Parsing origin: <c:/foo> against <http://example.org/foo/bar>
  assert_equals: origin expected "null" but got "file://"
[FAIL] Parsing origin: <h://.> against <about:blank>
  assert_equals: origin expected "null" but got "file://"
[FAIL] Parsing origin: <http://!"$&'()*+,-.;=_`{}~/> against <about:blank>
  assert_equals: origin expected "http://!\\"$&'()*+,-.;=_`{}~" but got "http://!\\"$&'()%2A+,-.;=_`{}~"
[FAIL] Parsing origin: <blob:ftp://host/path> against <about:blank>
  assert_equals: origin expected "null" but got "ftp://host"
[FAIL] Parsing origin: <blob:ws://example.org/> against <about:blank>
  assert_equals: origin expected "null" but got "ws://example.org"
[FAIL] Parsing origin: <blob:wss://example.org/> against <about:blank>
  assert_equals: origin expected "null" but got "wss://example.org"
[FAIL] Parsing origin: <wss://!"$&'()*+,-.;=_`{}~/> against <about:blank>
  assert_equals: origin expected "wss://!\\"$&'()*+,-.;=_`{}~" but got "wss://!\\"$&'()%2A+,-.;=_`{}~"
[FAIL] Parsing origin: <w://x:0> against <about:blank>
  assert_equals: origin expected "null" but got "file://"
[FAIL] Parsing origin: <http://example.com/U+d800ðŸ¾U+dfffï·ï·ï·¯ï·°ï¿¾ï¿¿?U+d800ðŸ¾U+dfffï·ï·ï·¯ï·°ï¿¾ï¿¿> against <about:blank>
  assert_equals: origin expected "http://example.com" but got ""
Harness: the test ran to completion.
